Split out util funcs
